% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_reservatorio.R
\name{add}
\alias{add}
\title{Dados dos reservatorios brasileiros}
\usage{
add(x, y)
}
\arguments{
\item{codigo_reservatorio}{Codigo do reservatorio a ser buscado.}

\item{data_inicial}{Data inicial do registro. Se for anterior a data de funcionamento, retorna a data de registro mais antiga.}

\item{data_final}{Data final do registro. Pode ser usado Sys.Date() para retornar a data de registro mais recente.}
}
\description{
Dados dos reservatorios brasileiros
}
\examples{
dados_reservatorio(19058, "2000-01-01", Sys.Date())

}
