% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_reservatorio.R
\name{dados_reservatorio}
\alias{dados_reservatorio}
\title{Dados dos reservatorios brasileiros}
\usage{
dados_reservatorio(codigo_reservatorio, data_inicial, data_final)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
dados_reservatorio()
}
\keyword{resevoir}
\keyword{water}
