% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_reservatorio_cantareira.R
\name{dados_reservatorio_cantareira}
\alias{dados_reservatorio_cantareira}
\title{Dados dos reservatorios brasileiros do Sistema Cantareira (Cantareira)}
\usage{
dados_reservatorio_cantareira(
  codigo_reservatorio_cantareira,
  data_inicial = "1980-01-01",
  data_final = Sys.Date()
)
}
\arguments{
\item{codigo_reservatorio_cantareira}{Codigo do reservatorio a ser buscado.}

\item{data_inicial}{Data inicial do registro. Se for anterior a data de funcionamento, retorna a data de registro mais antiga.}

\item{data_final}{Data final do registro. Pode ser usado Sys.Date() para retornar a data de registro mais recente.}
}
\description{
Dados dos reservatorios brasileiros do Sistema Cantareira (Cantareira)
}
\examples{
dados_reservatorio_cartareira(29001, "2000-01-01", Sys.Date())
}
