% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_reservatoriosBR.R
\name{info_reservatoriosBR}
\alias{info_reservatoriosBR}
\title{Informações das funções do pacote reservatoriosBR}
\usage{
info_reservatoriosBR()
}
\description{
Informações das funções do pacote reservatoriosBR
}
\examples{
info_reservatoriosBR()
}
