% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivational_message.R
\name{motivational_message}
\alias{motivational_message}
\title{motivational messages}
\usage{
motivational_message(msg)
}
\arguments{
\item{msg}{codigo para qual message quer mostrar.}
}
\description{
motivational messages
}
\keyword{internal}
