% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONS_EAR_subsistemas.R
\name{ONS_EAR_subsistemas}
\alias{ONS_EAR_subsistemas}
\title{Dados históricos de Energia Armazenada (EAR) por subsistema segundo a ONS}
\usage{
ONS_EAR_subsistemas(ano_inicial = 2000, ano_final = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{ano_inicial}{Ano inicial do registro. Série histórica tem início no ano 2000.}

\item{ano_final}{Ano final do registro. Pode deixar em branco para retornar a data de registro mais recente.}
}
\description{
Dados históricos de Energia Armazenada (EAR) por subsistema segundo a ONS
}
\examples{
ONS_EAR_subsistemas(ano_inicial=2000, ano_final=format(Sys.Date(), "\%Y"))
}
