% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONS_reservatorios.R
\name{ONS_reservatorios}
\alias{ONS_reservatorios}
\title{Dados atuais dos reservatorios brasileiros observados pela ONS}
\usage{
ONS_reservatorios(formato = "tabela")
}
\arguments{
\item{formato}{Formato dos dados a serem apresentados.}
}
\description{
Dados atuais dos reservatorios brasileiros observados pela ONS
}
\examples{
ONS_reservatorios(formato="tabela")
ONS_reservatorios(formato="resumo")
}
