% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservatorio_cantareira.R
\name{reservatorio_cartareira}
\alias{reservatorio_cartareira}
\title{Dados dos reservatorios brasileiros do Sistema Cantareira (Cantareira)}
\usage{
reservatorio_cartareira(
  codigo_reservatorio,
  data_inicial = "1980-01-01",
  data_final = Sys.Date()
)
}
\arguments{
\item{codigo_reservatorio}{Codigo do reservatorio a ser buscado, (voce pode consultar a tabela com os codigos chamando \code{tabela_reservatorios}).}

\item{data_inicial}{Data inicial do registro. Se for anterior a data de funcionamento, retorna a data de registro mais antiga.}

\item{data_final}{Data final do registro. Pode ser usado Sys.Date() para retornar a data de registro mais recente.}
}
\description{
Dados dos reservatorios brasileiros do Sistema Cantareira (Cantareira)
}
\examples{
reservatorio_cartareira(29001, "2000-01-01", Sys.Date())
}
