% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservatorio_nordeste_semiarido.R
\name{reservatorio_nordeste_semiarido}
\alias{reservatorio_nordeste_semiarido}
\title{Dados dos principais reservatorios brasileiros do modulo Nordeste e Semiarido do SAR}
\usage{
reservatorio_nordeste_semiarido(
  codigo_reservatorio,
  data_inicial = "1980-01-01",
  data_final = Sys.Date()
)
}
\arguments{
\item{codigo_reservatorio}{Codigo do reservatorio a ser buscado, (voce pode consultar a tabela com os codigos chamando \code{tabela_reservatorios}).}

\item{data_inicial}{Data inicial do registro. Se for anterior a data de funcionamento, retorna a data de registro mais antiga.}

\item{data_final}{Data final do registro. Pode ser usado Sys.Date() para retornar a data de registro mais recente.}
}
\description{
Dados dos principais reservatorios brasileiros do modulo Nordeste e Semiarido do SAR
}
\examples{
reservatorio_nordeste_semiarido(12354, "2012-09-10", Sys.Date())
}
