% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservatorios_ONS.R
\name{reservatorios_ONS}
\alias{reservatorios_ONS}
\title{Dados dos reservatorios brasileiros observados pela ONS}
\usage{
reservatorios_ONS(formato = "tabela")
}
\arguments{
\item{formato}{Formato do dados a serem apresentados.}
}
\description{
Dados dos reservatorios brasileiros observados pela ONS
}
\examples{
reservatorios_ONS(formato="tabela")
reservatorios_ONS(formato="resumo")
}
