% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_reservatorios.R
\name{tabela_reservatorios}
\alias{tabela_reservatorios}
\title{Tabela com dados dos reservatorios brasileiros}
\format{
Um dataframe com 805 observacoes e 12 variaveis
\describe{
\item{sistema}{sistema o qual o reservatorio pertence segundo o SAR}
\item{codigo}{o codigo do reservatorio segundo o SAR}
\item{reservatorio}{o nome do reservatorio}
\item{res_latitude}{latitude do reservatorio}
\item{res_longitude}{longitude do reservatorio}
\item{municipio}{municipio do reservatorio}
\item{codigo_municipio_ibge}{codigo do municipio na base do ibge}
\item{estado_nome}{estado em forma de nome}
\item{estado_sigla}{estado em forma de sigla}
\item{rio}{nome do rio}
\item{sub_bacia}{sub-bacia do rio}
\item{bacia}{bacia hidrográfica do rio}
}
}
\source{
ANA (Antônio Augusto, Diego Pena e Théo Albuquerque), editado por Bruno Mioto. obtido em 2021-08-22
}
\usage{
tabela_reservatorios()
}
\description{
Um dataset contendo os atributos dos principais reservatorios brasileiros.
}
\examples{
tabela_reservatorios()
}
