% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_reservatorios.R,
%   R/tabela_reservatorios_ONS.R
\name{tabela_reservatorios_ONS}
\alias{tabela_reservatorios_ONS}
\title{Tabela com dados dos reservatorios brasileiros}
\format{
Um dataframe com 708 observacoes e 5 variaveis
\describe{
\item{sistema}{sistema o qual o reservatorio pertence}
\item{estado_1}{estado o qual o reservatorio pertence}
\item{estado_2}{se o reservatorio fica entre estados, o segundo estado}
\item{reservatorio}{o nome do reservatorio}
\item{codigo}{o codigo do reservatorio segundo o SAR}
}

Um dataframe com 39 observacoes e 6 variaveis
\describe{

\item{reservatorio}{o nome do reservatorio}
\item{bacia}{bacia hidrológica a qual o reservatorio pertence}
\item{subsistema}{subsistema-ONS a qual o reservatorio pertence}
\item{estado_1}{estado o qual o reservatorio pertence}
\item{estado_2}{se o reservatorio fica entre estados, o segundo estado}
\item{codigo}{o codigo do reservatorio segundo o SAR}
}
}
\source{
\url{https://www.ana.gov.br/sar0/}. obtido em 2021-08-22

\url{http://www.ons.org.br/paginas/energia-agora/reservatorios}. obtido em 2021-08-30
}
\usage{
tabela_reservatorios_ONS()

tabela_reservatorios_ONS()
}
\description{
Um dataset contendo os atributos dos principais reservatorios brasileiros.

Um dataset contendo os atributos dos principais reservatorios brasileiros apresentados pela ONS.
}
\examples{
tabela_reservatorios()

tabela_reservatorios_ONS()
}
