% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultima_medicao.R
\name{ultima_medicao}
\alias{ultima_medicao}
\title{Ultima medicao dos reservatorios brasileiros}
\usage{
ultima_medicao(codigo_reservatorio)
}
\arguments{
\item{codigo_reservatorio}{Codigo do reservatorio a ser buscado.}
}
\description{
Ultima medicao dos reservatorios brasileiros
}
\examples{
ultima_medicao(19058)
}
