% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_reservatorios.R
\name{tabela_reservatorios}
\alias{tabela_reservatorios}
\title{Tabela com dados dos reservatorios brasileiros}
\format{
Um dataframe com 708 observacoes e 5 variaveis
\describe{
\item{sistema}{sistema o qual o reservatorio pertence}
\item{estado_1}{estado o qual o reservatorio pertence}
\item{estado_2}{se o reservatorio fica entre estados, o segundo estado}
\item{reservatorio}{o nome do reservatorio}
\item{codigo}{o codigo do reservatorio segundo o SAR}
}
}
\source{
\url{https://www.ana.gov.br/sar0/}. obtido em 2021-08-22
}
\usage{
tabela_reservatorios()
}
\description{
Um dataset contendo os atributos dos principais reservatorios brasileiros.
}
\examples{
tabela_reservatorios()

}
