% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_reservatorios.R
\docType{data}
\name{tabela_reservatorios}
\alias{tabela_reservatorios}
\title{Tabela com dados dos reservatorios brasileiros}
\format{
An object of class \code{data.frame} with 708 rows and 5 columns.
}
\usage{
data(tabela_reservatorios)
}
\description{
Tabela com dados dos reservatorios brasileiros
}
\examples{
data(tabela_reservatorios)
}
\keyword{datasets}
